refactor: app directory structure
