+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import CalculateScreen from './screens/CalculateScreen/CalculateScreen'
+import CalculateScreen from './screens/CalculateScreen'
 
 const App: React.FC = () => {
   return <CalculateScreen />
