+++ src/screens/CalculateScreen/CalculateScreen.tsx
@@ -4,9 +4,9 @@ import { useDispatch, useSelector } from 'react-redux'
 import { moneyActions, moneySelectors } from '../../store/money.slice'
 
 import { AppBar, Box, IconButton, Container, Stack, Typography, styled } from '@mui/material'
-import MoneyTotal from '../../components/MoneyTotal/MoneyTotal'
-import MoneyItems from '../../components/MoneyItems/MoneyItems'
-import SortMoneyItemsButton from '../../components/MoneyItems/components/SortMoneyItemsButton/SortMoneyItemsButton'
+import MoneyTotal from './components/MoneyTotal/MoneyTotal'
+import MoneyItems from './components/MoneyItems/MoneyItems'
+import SortMoneyItemsButton from './components/MoneyItems/components/SortMoneyItemsButton/SortMoneyItemsButton'
 import RestartAltIcon from '@mui/icons-material/RestartAlt'
 import useConfirmDialog from '../../components/ConfirmDialog/useConfirmDialog'
