+++ src/screens/CalculateScreen/components/MoneyItems/MoneyItems.styled.tsx
