+++ src/screens/CalculateScreen/components/MoneyItems/components/AddMoneyItem/AddMoneyItem.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react'
 import { useDispatch } from 'react-redux'
 import { Button } from '@mui/material'
 
-import { moneyActions } from '../../../../store/money.slice'
+import { moneyActions } from '../../../../../../store/money.slice'
 
 const AddMoneyItem: React.FC = () => {
   const dispatch = useDispatch()
