+++ src/screens/CalculateScreen/components/MoneyItems/components/MoneyItem/MoneyItem.styled.ts
