+++ src/screens/CalculateScreen/components/MoneyItems/components/MoneyItem/constants.ts
