+++ src/screens/CalculateScreen/components/MoneyItems/constants.ts
